Fixed a few warnings in the PISA target.

+++ target-pisa/syscall.c
@@ -1811,7 +1811,7 @@ sys_syscall(struct regs_t *regs,	/* registers to access */
 	/* copy target side I/O vector buffers to host memory */
 	for (i=0; i < /*iovcnt*/regs->regs_R[6]; i++)
 	  {
-	    iov[i].iov_base = (char *)MD_SWAPW((unsigned)iov[i].iov_base);
+	    iov[i].iov_base = (char *)MD_SWAPW((unsigned long)iov[i].iov_base);
 	    iov[i].iov_len = MD_SWAPW(iov[i].iov_len);
 	    if (iov[i].iov_base != NULL)
 	      {
